added seperation for browser tests
